mf-risk-assistant-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
# pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.27.0
# python-multipart==0.0.9
google-generativeai==0.3.1
fastmcp==0.1.0 